requirements.txt
opencv-python
python-chess
numpy
scikit-image
berserk